PYIC-866: Add command to produce public key JWKs
We need to be able to produce a JWK for a KMS key in a Concourse
pipeline. This adds a new command to allow us to do this.

It refactors out the existing logic into a separate class and adds it
and the new commands as sub-commands.
