+++ settings.gradle
@@ -1,2 +1,2 @@
-rootProject.name = 'di-ipv-kms-csr-generator'
+rootProject.name = 'di-ipv-kms-public-key-operations'
